+++ src/components/signForms/signInForm.tsx
@@ -1,0 +1,94 @@
+import Avatar from '@mui/material/Avatar';
+import Button from '@mui/material/Button';
+import CssBaseline from '@mui/material/CssBaseline';
+import TextField from '@mui/material/TextField';
+import Link from '@mui/material/Link';
+import Grid from '@mui/material/Grid';
+import Box from '@mui/material/Box';
+import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
+import Typography from '@mui/material/Typography';
+import Container from '@mui/material/Container';
+import { useAppDispatch, useAppSelector } from 'app/hooks';
+import { useNavigate } from 'react-router-dom';
+import { logIn } from 'components/header/authSlice';
+
+export default function SignInForm() {
+  const dispatch = useAppDispatch();
+  const language = useAppSelector((state) => state.lang.current);
+  const navigate = useNavigate();
+
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    const data = new FormData(event.currentTarget);
+    dispatch(logIn());
+    console.log({
+      email: data.get('email'),
+      password: data.get('password'),
+    });
+  };
+
+  return (
+    <Container component="main" maxWidth="xs">
+      <CssBaseline />
+      <Box
+        sx={{
+          marginTop: 4,
+          marginButton: 2,
+          display: 'flex',
+          flexDirection: 'column',
+          alignItems: 'center',
+        }}
+      >
+        <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+          <LockOutlinedIcon />
+        </Avatar>
+        <Typography component="h1" variant="h5">
+          {language === 'EN' ? 'Sign In' : 'Вход'}
+        </Typography>
+        <Box component="form" onSubmit={handleSubmit} noValidate sx={{ mt: 1 }}>
+          <TextField
+            margin="normal"
+            required
+            fullWidth
+            id="login"
+            label={language === 'EN' ? 'Login' : 'Логин'}
+            type="text"
+            name="login"
+            autoComplete="username"
+            autoFocus
+          />
+          <TextField
+            margin="normal"
+            required
+            fullWidth
+            name="password"
+            label={language === 'EN' ? 'Password' : 'Пароль'}
+            type="password"
+            id="password"
+            autoComplete="current-password"
+          />
+          <Button type="submit" fullWidth variant="contained" sx={{ mt: 3, mb: 2 }}>
+            {language === 'EN' ? 'Sign In' : 'Войти'}
+          </Button>
+          <Grid container>
+            <Grid item xs></Grid>
+            <Grid item>
+              <Link
+                href="/signUp"
+                onClick={(e) => {
+                  e.preventDefault();
+                  navigate('/signUp');
+                }}
+                variant="body2"
+              >
+                {language === 'EN'
+                  ? `Don't have an account? Sign Up`
+                  : 'Нет аккаунта? Зарегестрируйтесь'}
+              </Link>
+            </Grid>
+          </Grid>
+        </Box>
+      </Box>
+    </Container>
+  );
+}
