+++ src/components/signForms/signUpForm.tsx
@@ -1,0 +1,118 @@
+import * as React from 'react';
+import Avatar from '@mui/material/Avatar';
+import Button from '@mui/material/Button';
+import CssBaseline from '@mui/material/CssBaseline';
+import TextField from '@mui/material/TextField';
+import Link from '@mui/material/Link';
+import Grid from '@mui/material/Grid';
+import Box from '@mui/material/Box';
+import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
+import Typography from '@mui/material/Typography';
+import Container from '@mui/material/Container';
+import { useAppDispatch, useAppSelector } from 'app/hooks';
+import { useNavigate } from 'react-router-dom';
+import { logIn } from 'components/header/authSlice';
+
+export default function SignUpForm() {
+  const dispatch = useAppDispatch();
+  const language = useAppSelector((state) => state.lang.current);
+  const navigate = useNavigate();
+
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    const data = new FormData(event.currentTarget);
+    dispatch(logIn());
+    console.log({
+      email: data.get('email'),
+      password: data.get('password'),
+    });
+  };
+
+  return (
+    <Container component="main" maxWidth="xs">
+      <CssBaseline />
+      <Box
+        sx={{
+          marginTop: 4,
+          marginButton: 2,
+          display: 'flex',
+          flexDirection: 'column',
+          alignItems: 'center',
+        }}
+      >
+        <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+          <LockOutlinedIcon />
+        </Avatar>
+        <Typography component="h1" variant="h5">
+          {language === 'EN' ? 'Sign Up' : 'Регистрация'}
+        </Typography>
+        <Box component="form" noValidate onSubmit={handleSubmit} sx={{ mt: 3 }}>
+          <Grid container spacing={2}>
+            <Grid item xs={12} sm={6}>
+              <TextField
+                autoComplete="given-name"
+                name="firstName"
+                required
+                fullWidth
+                id="firstName"
+                label={language === 'EN' ? 'First Name' : 'Имя'}
+                autoFocus
+              />
+            </Grid>
+            <Grid item xs={12} sm={6}>
+              <TextField
+                required
+                fullWidth
+                id="lastName"
+                label={language === 'EN' ? 'Last Name' : 'Фамилия'}
+                name="lastName"
+                autoComplete="family-name"
+              />
+            </Grid>
+            <Grid item xs={12}>
+              <TextField
+                required
+                fullWidth
+                id="login"
+                label={language === 'EN' ? 'Login' : 'Логин'}
+                type="text"
+                name="login"
+                autoComplete="username"
+              />
+            </Grid>
+            <Grid item xs={12}>
+              <TextField
+                required
+                fullWidth
+                name="password"
+                label={language === 'EN' ? 'Password' : 'Пароль'}
+                type="password"
+                id="password"
+                autoComplete="new-password"
+              />
+            </Grid>
+          </Grid>
+          <Button type="submit" fullWidth variant="contained" sx={{ mt: 3, mb: 2 }}>
+            {language === 'EN' ? 'Sign Up' : 'Зарегестрироваться'}
+          </Button>
+          <Grid container justifyContent="flex-end">
+            <Grid item>
+              <Link
+                href="/signIn"
+                onClick={(e) => {
+                  e.preventDefault();
+                  navigate('/signIn');
+                }}
+                variant="body2"
+              >
+                {language === 'EN'
+                  ? 'Already have an account? Sign in'
+                  : 'Уже есть аккаунт? Войдите'}
+              </Link>
+            </Grid>
+          </Grid>
+        </Box>
+      </Box>
+    </Container>
+  );
+}
